:zap: Introduce `EditorToolbar` component that is a lazy wrapper around editor controls. Also lazy load Plain and Rich text editors.
