--- src/lib/components/EditorToolbarLazy.svelte
@@ -1,7 +1,0 @@
-<script>
-  export let currentUser;
-</script>
-
-{#await import('$lib/components/EditorToolbar.svelte') then EditorToolbar}
-  <EditorToolbar.default {currentUser} on:cancel on:save />
-{/await}
