+++ src/routes/blog/new/+page.svelte
@@ -6,7 +6,7 @@
   import EditableWebsiteTeaser from '$lib/components/EditableWebsiteTeaser.svelte';
   import Article from '$lib/components/Article.svelte';
   import NotEditable from '$lib/components/NotEditable.svelte';
-  import EditorToolbarLazy from '$lib/components/EditorToolbarLazy.svelte';
+  import EditorToolbar from '$lib/components/EditorToolbar.svelte';
 
   export let data;
 
@@ -45,7 +45,7 @@
 </svelte:head>
 
 {#if editable}
-  <EditorToolbarLazy {currentUser} on:cancel={discardDraft} on:save={createArticle} />
+  <EditorToolbar {currentUser} on:cancel={discardDraft} on:save={createArticle} />
 {/if}
 
 <WebsiteNav bind:editable bind:showUserMenu {currentUser} />
